Extract SqlFamily to its own module
