+++ libs/sql-connection/src/lib.rs
@@ -6,13 +6,15 @@ mod connection_info;
 mod generic_sql_connection;
 mod mysql;
 mod postgres;
+mod sql_family;
 mod sqlite;
 mod traits;
 
 pub use connection_info::ConnectionInfo;
-pub use generic_sql_connection::{GenericSqlConnection, SqlFamily, SUPPORTED_SCHEMES};
+pub use generic_sql_connection::GenericSqlConnection;
 pub use mysql::*;
 pub use postgres::*;
+pub use sql_family::SqlFamily;
 pub use sqlite::*;
 pub use traits::*;
